feat: event editing support to society event form
Recurrence controls are hidden when editing, and the form is pre-populated with event data.
The submit button label and summary view are updated to reflect edit mode.
Event update logic is added to the submission handler.
Currently only a single event can be edited at a time. To be able to mass edit events, we'll need a hash that identifies the events that are linked together